Made states un-register all buttons before switching
Previously, buttons were never getting removed from the FlxMouseEventManager
object. This could probably cause a bunch of problems, including causing a crash
if buttons in the two states overlap and the mouse is hovering over the overlap point
when the state is switched. This happened because the mouse manager was
trying to access a field of a button that no longer existed.

This commit makes the states responsible for un-registering all buttons before
switching. I also tried making the buttons responsible for un-registering themselves
when killed, but it didn't solve the issue.