Reorganized files and made Button callback arguments optional.
The project's files were moved around to hopefully make the file organization
stay relatively clean as the game grows in complexity. Code-wise, the changes
to make this happen should only be in package and import statements.

Button's callback arguments are now optional, allowing calls to the constructor
to skip them or enter 'null' if one of the later argument actually needs to do
something. To achieve this, a the EmptyFunction class was made, which
lets me call a short-named function instead of typing out an inline function
every time I want a placeholder function that doesn't do anything.