Reworked the design of Button and added documentation.
After thinking over the overall design of the project, particularly the
classes related to buttons, I decided to change up the design a bit.

Instead of passing in a bunch of parameters every time a Button is constructed,
I pass in the class that contains static fields describing that type of button's functionality.
This cuts down on the size of the button constructor pretty heavily, and also means I don't
have to worry about optional arguments or trying to do something weird with a separate
factory class.

Since there's just one example button type right now, I had to get rid of the other buttons on
the main menu. Now, however, the groundwork is laid for future button logic-defining,
and the whole process should be pretty smooth.