Created new GameManager class.
This should be used to facilitate switching between states and tracking general
game information. Right now it seems like this will be useful in the future. If it turns
out that it's not really needed, I can distribute its functions among the other classes
in the game.

I think this will eventually be set up as a plugin in FlxG, but I'll have to do some additional
research to figure out how that will work. For now, all I'm using it for is static functions,
so this is fine.