Expanded Button and "tab" classes with new logic.
These now move to the side when hovered over, and move back to their 'anchor"
position when the mouse leaves. Clicking on a tab makes it active and deactivates
the others.

With this commit, I started to realize the drawbacks of this static class approach.
It's really cool if all of my classes have distinct behaviors, but since static fields
aren't inherited, there isn't a convenient way to combine common behavior between classes.