Added MovingButton class.
Adds back in the rest of the functionality that the old Button class had,
and also pulls some of the functionality from the old tab classes.
If there's ever a button that needs to shift between positions when
active or inactive, this class should be pretty handy.

Had to do some casting shenanigans In the mActivate and mDeactive
functions, since the functions they overload don't take a specific enough
type as a parameter. To work around it, casted the ActiveButton parameter
into a MovingButton.