Updated states to use the new style of button classes.
The only real hiccup in this commit was that I can no longer
grab the width and height of a particular button without instantiating
an instance of it, which (unfortunately) is tied to registering that button
with the FlxMouseEventManager.

I may go back and add a static members of classes when I need
their dimensions, but that feels kinda messy, not that the current
solution is any cleaner.