No longer use button parameter when not necessary.
Previously, I would use the "button" parameter passed into some non-static member
functions of button template classes when the "button" and the instance that was calling
the method were the same thing.

Ends up cleaning up the code a bit and make it clear that the "button" parameter doesn't
really matter at all in the non-static functions.