Reworked MovingButton and Tab functionality.
The old logic used in the MovingButton class wasn't pretty hard to interpret, I think.
The new setup has variables that track how much the anchor and offset should change
whenever the button is activated/deactivated.

It also separates the execution of button movement from the default behavior of a moving
button, which is more in the spirit of the button template, I think. It's up to the child classes
when to move the button, the MovingButton class just contains the tools to get it done.