Added Event.hx
This file serves a couple of purposes. It defines the "Event" type, which
is effectively the same as an Int. The idea is that it holds data that is smaller
than an integer, which is accessible via bitwise operations. The file also
declares the class EventExtender, which is what should make Event a cool
thing to work with.

Any file can add the line "using utilities.Event" and then any event should be
able to use the methods described within, according to the rules of Haxe's
Static Extension.