Effectively made Tab function as a child of Subject.
Accomplished using tink_lang's Syntactic Delegation feature.
It sets up functions in the tab class so that calling myTab.notify()
is the same as calling myTab.sub.notify().

This is nice when an observer is trying to add itself as a listener
to the subject, because Tab gets to behave like it inherited Subject's
variables and methods, even though it really didn't. Tink_lang was
necessary because Haxe does not support multiple inheritance.