Joined ShopState to tab buttons via Observer pattern.
This is kind of a big step, I think. It demonstrates the desired
sort of relationship between a button-managing entity and its
buttons. Previously I had the tabs just tightly coupled to the
shopstate class, but now they act in the manner that I orignally
wanted them to.

A similar sort of structure will be used when building the
brew ingredients table, which is the overall goal of this branch.