Added Hideable class to take hide/reveal logic out of Button.
This felt necessary because there are other types of objects (besides children of Button)
that will need to be hidden/revealed whenever the player switches between different sections
of the shop.

Under the new system, anything that needs to be "hideable" will make sure that Hideable is
somewhere in its inheritance tree, preferrably near the top. Any behavior outside of the basic
"can't be seen, doesn't update" stuff can be handled by overriding hide and reveal in the child
classes, as Button does.