Added IngredientTable class.
IngredientTable will be responsible for creating and observing a variety of buttons.
In this iteration, it handles the creation of IngredientHexes and associating some
(dummy) data with them.

Future commits will set up a system to associate useful data with the different
buttons, add additional button types the table contains, add a background graphic,
and more!