Created DisplaySprite class
This class should be the parent of all objects in the game that
need to move and animate, but have no interaction component
or containing duties.

At the moment, it creates single-frame animations labelled according
to the image that the animation will display. This is so one sprite can
display itself as a wide variety of different images (which is needed
for ingredient graphic displaying).