Created ColorArray class.
As the documentation comment at the top of the file says, this class
is a proxy for an ordinary Haxe array. It is needed because HaxeFlixel's
FlxBar's data binding-esque feature requires an object's field to be specified,
and a normal Haxe Array can't be accessed in that manner.

What this class does is set up properties for each of the different ingredient
colors, and specifies that they all have a getter function. The getter functions
simply get the appropriate index from "array". This means that any attempt to
access "aColorArray.red" actually calls the get_red() function specified in the body.