Updated hover & release notify callbacks.
This includes creating separate functions for updating
HoverBars and LockedBars. It felt kinda messy to have
all of the loop business inside the set*Info() functions.

This also ensures that the see-through hover bars are
properly updated whever a new button is hovered over or
a button is released over.