Added NO_EVENT & NO_TYPE enum values.
EventData and ButtonTypes felt like they should specify some
invalid value that can be referenced when checking for events
with unset data.

This "invalid value" may be better off in some different place than
these enums, though. If it were, I may be able to find some way
to reference the number of entries in the enum for determining
things like the # of different valid button types.