Reworked how notified events are handled.
Rather than keep using a hard-to-read system of callbacks
that required knowledge of the FlxMouseEventManager's
inner workings to understand, I decided to simplify things.
I did this by separating the gathering of event information
from the process of acting upon that information. The explanation
below should make it clear.

Previously, the IngredientTable would immediately respond to any
event that was sent to it via the subject-observer notification system.
This meant that any events recieved from buttons would be resolved
in the same order they arrived, which was determined by the inner
workins of the FlxMouseEventManager. Specifically, the mouse
events happened in this order:

MouseOver
MouseOut
MouseDown/Up (only one can happen on a frame, I believe)

This ended up making IngredientTable's logic more complicated,
since I wanted to clear the currently displayed data after a mouseOut
and display the new button's info after a mouseOver. If I used a naieve
approach under the old system, problems occurred when the mouse
left one button and entered another on the same frame: I would display
the new data first, then clear that new data before the user had a chance
to se it. I tried to fix this problem with a somewhat convoluted set of
logical tests.

The new system is much cleaner. Events are stored in an array of events
when they arrive (indexed by the type of event) and are later handeld within
the update function. There they are always handled in the same order:

MouseOut
MouseOver
MouseDown
MouseUp

This should make the logic more understandable for everyone!