Added colorChanged var to IngredientTable.
This is a positive change because it lets me reduce how
often the updatePotionData function will need to be called
in the update() function. Without this, I would have to call
updatePotionData every frame, but now it'll only need to be
called when some change has happened to the hovered
or selected values!