Created BrewContainer class.
This class is designed to contain all parts of the ShopState's Brew
section, since the ingredient table doesn't quite hold everything
the state needs inside of it.

It also handles some communication between the IngredientTable
and buttons that exist outside of it, which is why a few function
definitions needed to be added to IngredientTable.