Created data file for potion effects.
This file is similar in function to IngredientData. It will
be read and parsed as JSON, then put in a data structure
where it can be accessed to get the potion effect names.

The effect names in the file are placeholders, and should
eventually be replaced with potion effects like healing,
fire-breathing, etc.