Use PotionEffects.txt for potion effect text.
Previously, potion effects were generated within
PotionData's updateText function. This was pretty
limited, because other parts of the program that need
potion effects (such as in customer generation) couldn't
get to it, and the names generated would either have to
be fairly simple or hard-coded in somewhere.

Now the effects are specified in JSON in PotionEffects.txt,
so they can be accessed by anything that needs it, including
PotionData.