Added numbers to the bar display.
This makes it very clear how many color "points"
the current potion has. The numbers also turn
red/green to show that numbers are increasing/
decreasing. Previously it was ambiguous whether
see-through bars were representing a gain or a loss
in color values, but that isn't a problem any more.