Hideable name changed to AdvancedSprite.
Since many of the components of the game need to undergo some
sort of transition when the player chooses to advance time, I decided
to add another overridable function/static function pair to Hideable.

Since Hideable is now serving two different purposes, a name change
seemed appropriate, and I picked AdvancedSprite to mirror the game's
existing AdvancedState class.

The AdvanceTimeReset static function is similar to the Hide and Reveal
functions; it will be used with FlxGroup's forEach function so it can be
applied to large groups of objects at once.