Changed CustomerCard functionality.
Previously, clicking on CustomerCard would open up a substate
with options to either quit the game or close the substate.

Now, it functions much like the Cauldron class, used in BrewState.
When activated it is in a permanent "hovered" state until deactivation.
It fires off notifications when it recieves a mouseUp event, etc.