Added CustomerCover & Lock classes.
These are very similar in purpose and function to the LockButton
and LockCover classes used to lock the Ingredient table after
clicking the brew button, but are instead intended for customer
cards.

I'm not super comfortable with this strategy being used again, but
it works and time left in the project is running out. It would be better
to have some sort of template class for both lock & cover classes to
inherit from, but that's something that'll have to be put in the post-school
project to-do list. Unless this strategy is trashed entirely, which is very
possible.