Integrated CustomerLock with CustomerCard.
This adds the needed functions for CustomerCard to
manage and use a CustomerLock (and, indirectly, a
CustomerCover).

I guess this situation already existed before, but I
just noticed that CustomerCard now inhabits some wierd
space between container and Button, which will only get
more pronounced as development continues. I don't think
that this is a problem, but it may indicate that I need a new
category to put these container/buttons.