Hid IngredientTable's lockButton during init.
Not technically needed, IngredientTable and everything
else in BrewContainer begin hidden. A weird situation can
occur if the lock isn't hidden from the outset, though, so
putting this here safeguards against that happening if
the game changes so you could start at the ingredientTable
in the future (like in a tutorial, perhaps).