Created CustContainer class.
Similar to BrewContainer in size and scope. It contains
all buttons related to the customer section of the ShopState
and responds to their button events.

It also passes along if the "next phase button" was clicked,
since it doesn't have the knowledge of the state as a whole to
advance everything in the shop to the next time section.