Created PotionRow class.
Fairly straightforward container class that holds a
set of 9 buttons InvPotionButtons. It has two different
graphics to display, one if a button in its row is selected
and the other if no button in its row is selected. Actually
changing between the two different modes requires a
external class (one that is listening to its contents) to call
the proper function.