Moved potion array job to GameManager.
In a past commit, I added potion-generating functionality to
InventoryTable for testing purposes. In this commit, I move that
behavior over to its likely long-term home: the GameManager class.

Other classes will need to interact with the potion array in limited ways,
(potions needed to be generated based off of brew table information
and then stored in the array, for example) so it seemed best to put this
data in a single class that everything else could reach.

I know that singleton's aren't great for a lot of reasons, but I think the
GameManager will end up being useful for this sort of game.