InvTable clears selected potion info/buttons when hidden.
Previously, the game would remember which button the player had selected
even if the player navigated away from the inventory table. That presents a
problem when considering the fact that potions can (eventually) be removed
from the inventory table. What happens if the player removes a potion then goes
back to the inventory table? It's possible that an empty button would be selected,
which would not be good.