Added button-updating & button-disabling to InventoryTable.
The big upshot of this is that InventoryTable can now properly display
a dynamically changing set of potions. If there are fewer potions in the
array than will fit in the table, the unused buttons are deactivated in the
FlxMouseEventManager, meaning that they won't trigger any mouse
events. If there are too many potions some won't be displayed, which is
less than ideal. However, some groundwork has beeen laid for a multi-
page InventoryTable!