Fixed potion display bug.
If the player moved the mouse from one button to another in a
single frame, it would not the display text at the bottom with the
newly hovered-button's text, because it didn't have an if statement
properly set up to address the situation of moving from one hover
button to another. I believe this fixes the problem.