Implemented potion-selling substate.
This substate is triggered by selecting a customer card
and pressing the check mark button. The substate displays
the potion inventory, allowing the user to either select and sell
a potion to the customer or cancel the action.

This feature isn't fully complete (it still needs to filter out potions
that this particular customer does not want) but it's a solid step
toward completing the project.