Made InvPotionButtons deactivateable.
Rather than use existing "lock cover" logic for selectively deactivating
each inventory button, I put together a new, nicer strategy for deactivating
buttons. The GrayOutable interface specifies the components a button
needs so that they can be "grayed out" and deactivated.

One of the big changes that's needed for the potions is that each button
needs to have access to any other images that should be tinted gray. This
has been done here with the InvPotionButton.